+++ src/common/settings.ts
@@ -55,6 +55,8 @@ export class Chat {
     public captureTalk: boolean = false;
     //list of lines to capture
     public lines: string[] = [];
+    //dont capture when window hidden
+    public noCaptureHidden: boolean = false;
     public alwaysOnTop: boolean = false;
     public alwaysOnTopClient: boolean = true;
     public log: boolean = false;
@@ -138,7 +140,8 @@ export let SettingList: any[] = [
     ['chat.captureReviews', 0, 1, false],
     ['chat.captureTells', 0, 1, false],
     ['chat.captureTalk', 0, 1, false],
-    ['chat.gag', 0, 1, false]
+    ['chat.gag', 0, 1, false],
+    ['chat.noCaptureHidden', 0, 1, false]
 ];
 
 export class Settings {